Add stm32f72b-disco uart1 example with config
This commit contains a suggestion for adding a Config associated type to
the Board trait.

The motivation for this addition is that while writing a UART example
specific for this board, I wanted to be able to configure/specify the
Port/Pins that USART can use (in this specific case USART1).